+++ internal/provider/resource_pullzone.go
@@ -1320,6 +1320,7 @@ func (r *PullzoneResource) Schema(ctx context.Context, req resource.SchemaReques
 
 func (r *PullzoneResource) ConfigValidators(ctx context.Context) []resource.ConfigValidator {
 	return []resource.ConfigValidator{
+		pullzoneresourcevalidator.OriginComputeScript(),
 		pullzoneresourcevalidator.MiddlewareScript(),
 		pullzoneresourcevalidator.PermacacheCacheExpirationTime(),
 		pullzoneresourcevalidator.CacheStaleBackgroundUpdate(),